+++ zuluCrypt-cli/socket/socket.h
@@ -59,9 +59,9 @@ int SocketConnect( socket_t ) ;
 
 int SocketListen( socket_t ) ;
 
-int SocketGetData( socket_t,char ** buffer,size_t len ) ;
+size_t SocketGetData( socket_t,char ** buffer,size_t len ) ;
 
-int SocketSendData( socket_t,const char * buffer,size_t len ) ;
+ssize_t SocketSendData( socket_t,const char * buffer,size_t len ) ;
 
 void SockectSetReadOption( socket_t,int option ) ;
